requirements.txt
AHRS==0.3.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
Flask==3.0.0
fonttools==4.46.0
future==0.18.3
idna==3.6
iso8601==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
pyparsing==3.1.1
pyserial==3.5
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.4
six==1.16.0
urllib3==2.1.0
Werkzeug==3.0.1
